# Issue #37: Fix Dashboard Scroll Performance with Virtual Scrolling