--- a/frontend/package.json
+++ b/frontend/package.json
@@ -25,6 +25,7 @@
     "@supabase/supabase-js": "^2.57.2",
     "@tanstack/react-query": "^5.87.1",
     "@tanstack/react-query-devtools": "^5.87.1",
+    "@types/react-window": "^1.8.8",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
@@ -34,6 +35,8 @@
     "next-themes": "^0.4.6",
     "react": "19.1.0",
     "react-dom": "19.1.0",
+    "react-window": "^2.1.0",
+    "react-window-infinite-loader": "^1.0.10",
     "sonner": "^2.0.7",
     "tailwind-merge": "^3.3.1"
   },