--- a/frontend/src/app/dashboard/page.tsx
+++ b/frontend/src/app/dashboard/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useMemo, useCallback } from 'react';
 import { Header } from '../../components/Header';
 import { RepositoryCard } from '../../presentation/components/RepositoryCard';
 import { WorkspacePanel } from '../../presentation/components/WorkspacePanel';
@@ -50,18 +50,18 @@ export default function Dashboard() {
     firstRepo: repositories?.[0]?.name
   });
 
-  // Handle repository selection
-  const handleRepositorySelect = (repository: Repository) => {
+  // Handle repository selection - memoized to prevent re-renders
+  const handleRepositorySelect = useCallback((repository: Repository) => {
     setSelectedRepository(repository);
-  };
+  }, []);
 
-  // Handle closing workspace
-  const handleCloseWorkspace = () => {
+  // Handle closing workspace - memoized
+  const handleCloseWorkspace = useCallback(() => {
     setSelectedRepository(null);
-  };
+  }, []);
 
-  // Handle repository connection
-  const handleRepositoryConnect = async (repositoryId: number) => {
+  // Handle repository connection - memoized
+  const handleRepositoryConnect = useCallback(async (repositoryId: number) => {
     console.log('Connecting repository:', repositoryId);
     try {
       await connectRepository(repositoryId);
@@ -69,22 +69,26 @@ export default function Dashboard() {
     } catch (error) {
       console.error('Failed to connect repository:', error);
     }
-  };
+  }, [connectRepository]);
 
-  // Filter repositories based on search and filters
-  const filteredRepositories = repositories?.filter((repo) => {
-    const matchesSearch = !searchQuery || 
-      repo.name.toLowerCase().includes(searchQuery.toLowerCase()) ||
-      repo.description?.toLowerCase().includes(searchQuery.toLowerCase());
+  // Filter repositories based on search and filters - memoized for performance
+  const filteredRepositories = useMemo(() => {
+    if (!repositories) return [];
     
-    const matchesLanguage = !languageFilter || repo.language === languageFilter;
-    
-    const matchesConnection = connectionFilter === 'all' || 
-      (connectionFilter === 'connected' && repo.is_connected) ||
-      (connectionFilter === 'disconnected' && !repo.is_connected);
-    
-    return matchesSearch && matchesLanguage && matchesConnection;
-  }) || [];
+    return repositories.filter((repo) => {
+      const matchesSearch = !searchQuery || 
+        repo.name.toLowerCase().includes(searchQuery.toLowerCase()) ||
+        repo.description?.toLowerCase().includes(searchQuery.toLowerCase());
+      
+      const matchesLanguage = !languageFilter || repo.language === languageFilter;
+      
+      const matchesConnection = connectionFilter === 'all' || 
+        (connectionFilter === 'connected' && repo.is_connected) ||
+        (connectionFilter === 'disconnected' && !repo.is_connected);
+      
+      return matchesSearch && matchesLanguage && matchesConnection;
+    });
+  }, [repositories, searchQuery, languageFilter, connectionFilter]);
   
   console.log('🔍 Filter results:', {
     totalRepos: repositories?.length || 0,
@@ -95,13 +99,17 @@ export default function Dashboard() {
     firstFilteredRepo: filteredRepositories[0]?.name
   });
 
-  // Get unique languages for filter
-  const availableLanguages = Array.from(
-    new Set(repositories?.map(repo => repo.language).filter(Boolean) || [])
-  ).sort();
+  // Get unique languages for filter - memoized for performance
+  const availableLanguages = useMemo(() => {
+    if (!repositories) return [];
+    
+    return Array.from(
+      new Set(repositories.map(repo => repo.language).filter(Boolean))
+    ).sort();
+  }, [repositories]);
 
-  // Handle retry for different error scenarios
-  const handleRetry = () => {
+  // Handle retry for different error scenarios - memoized
+  const handleRetry = useCallback(() => {
     if (!isOnline) {
       // For offline scenarios, just show a message about connectivity
       alert('Please check your internet connection and try again.');
@@ -109,7 +117,7 @@ export default function Dashboard() {
     }
     
     refetchRepositories();
-  };
+  }, [isOnline, refetchRepositories]);
 
   // Show workspace if a repository is selected
   if (selectedRepository) {