--- a/frontend/src/app/layout.tsx
+++ b/frontend/src/app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import { QueryProvider } from "../providers/QueryProvider";
 import { AuthProvider } from "@/contexts/AuthContext";
+import { ThemeProvider } from "../providers/ThemeProvider";
 import { ErrorBoundary } from "../components/ErrorBoundary";
 import "./globals.css";
 
@@ -47,13 +48,20 @@ export default function RootLayout({
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
         <ErrorBoundary level="critical" showDetails={process.env.NODE_ENV === 'development'}>
-          <AuthProvider>
-            <QueryProvider>
-              <ErrorBoundary level="page">
-                {children}
-              </ErrorBoundary>
-            </QueryProvider>
-          </AuthProvider>
+          <ThemeProvider
+            attribute="class"
+            defaultTheme="system"
+            enableSystem
+            disableTransitionOnChange={false}
+          >
+            <AuthProvider>
+              <QueryProvider>
+                <ErrorBoundary level="page">
+                  {children}
+                </ErrorBoundary>
+              </QueryProvider>
+            </AuthProvider>
+          </ThemeProvider>
         </ErrorBoundary>
       </body>
     </html>