--- a/frontend/src/components/ThemeToggle.tsx
+++ b/frontend/src/components/ThemeToggle.tsx
@@ -1,0 +1,71 @@
+'use client';
+
+import { useTheme } from 'next-themes';
+import { useEffect, useState } from 'react';
+import { Sun, Moon, Monitor } from 'lucide-react';
+import { Button } from './ui/button';
+
+export function ThemeToggle() {
+  const [mounted, setMounted] = useState(false);
+  const { theme, setTheme } = useTheme();
+
+  useEffect(() => {
+    setMounted(true);
+  }, []);
+
+  if (!mounted) {
+    return (
+      <Button variant="ghost" size="sm" className="h-9 w-9 px-0">
+        <Monitor className="h-4 w-4" />
+        <span className="sr-only">Toggle theme</span>
+      </Button>
+    );
+  }
+
+  const cycleTheme = () => {
+    if (theme === 'light') {
+      setTheme('dark');
+    } else if (theme === 'dark') {
+      setTheme('system');
+    } else {
+      setTheme('light');
+    }
+  };
+
+  const getIcon = () => {
+    switch (theme) {
+      case 'light':
+        return <Sun className="h-4 w-4" />;
+      case 'dark':
+        return <Moon className="h-4 w-4" />;
+      default:
+        return <Monitor className="h-4 w-4" />;
+    }
+  };
+
+  const getLabel = () => {
+    switch (theme) {
+      case 'light':
+        return 'Switch to dark mode';
+      case 'dark':
+        return 'Switch to system theme';
+      default:
+        return 'Switch to light mode';
+    }
+  };
+
+  return (
+    <Button
+      variant="ghost"
+      size="sm"
+      onClick={cycleTheme}
+      className="h-9 w-9 px-0 transition-all duration-200 hover:bg-gray-100 dark:hover:bg-gray-800"
+      aria-label={getLabel()}
+    >
+      <div className="transition-transform duration-200 hover:scale-110">
+        {getIcon()}
+      </div>
+      <span className="sr-only">{getLabel()}</span>
+    </Button>
+  );
+}