… PERFORMANCE_REPORT.md
… frontend/performance-results-2025-09-09T06-25-31.131Z.json
… frontend/run-performance-test.js
… frontend/src/__tests__/hooks/useIntersectionObserver.test.ts
… frontend/src/hooks/useIntersectionObserver.ts
… frontend/src/presentation/components/VirtualRepositoryGrid.tsx
… frontend/test-data-generator.js
… performance-test.js
… virtual-scrolling-implementation.md